fix: Do not include file type in docsets when they are set explicitly
This change will stop vim-dasht from including the name of the file type
in the docsets, because this leads to matches in irrelevant docsets being
included in the output.

For example, if the file type is `c`, *all* docsets that have a `c` in
its name will be searched.

The user cannot change this behavior from the config, even when setting
`g:dasht_filetype_docsets`:

    let g:dasht_filetype_docsets = {}
    let g:dasht_filetype_docsets['java'] = ['Java_SE11', 'JUnit5']

If the user has a `javascript` docset installed, it will still be
used for `java` files despite the configuration.

This changes this behavior to always use *only* those docsets that have
been specified explicitly by the user. If there is no configuration for
the current file type, it falls back to the old behavior and searches in
all docsets that have the file type name in their name.